feat: add tracing, refactor
